requirement.txt
request
pillow
pytesseract
tesseract-ocr
tesseract
selenium